Add RFC2822 date header
